Create 160-A twins
In this problem i am finding the sum of the array and then sorting the array and after that i am again finding the sum from the n-1th index to 0th index .If new sum becomes greater than the old sum/2 then program comes out of the loop and we are printing the value of c(no. of coins).